Small Bug Fix - PowerRule called without 'new' operator
In CalcFunctions (an array of functions), the PowerRule function wasn't
called with a 'new' operator. It failed gracefully, but the variables
wouldn't be used (specifically, the variable name, if it wasn't 'x'
(i.e. 'y')wouldn't show up correctly). This commit fixes the problem
both ways: by adding the 'new' operator and also having the variables
come along for the graceful fail.