Add differentiation capability to KAS
Add function 'differentiate' to all nodes (e.g. sum rule, product rule,
power rule, trig derivatives, log derivative) that returns an
expression (which you can then tex or differentiate again).

Add four helper functions:
- sameAs(other): checks if they have the same arguments.
- copy(): shallow copy of arguments (used primarily to copy unknown
expressions)
- contains(elem): (Expr) determines if any node in expression is
sameAs(elem)
- decrement(): returns a Num  decremented

One concern is that the time to 'simplify' a derivative may be on the
order of 100ms due to multiplicity of ways an expression can go through
factor/collect/expand functions. An option to semi-simplify (get rid of
zeros and ones) might be helpful here, especially so that the
derivative is recognizable when paired with the original expression.

The test.html is copy/pasted from the KAS project, with the addition of
test cases for 'differentiate'. Some test cases (outside of
'differentiate' fail due to lack of support for the unicode minus sign)