Add parseArray
The function 'parseArray' takes a (nested) array [operator, args] and
parses it into an Expr. In this commit, there is only support for Int,
Float, Const (e and pi), Var, Add, Mul, Pow, Log, and Trig. (That is,
there is no Func, Eq, Const (i), Abs). Furthermore, it'll break when
given an argument like 'color'.

This is an improvement (time, simplicity) over converting the array
into text and then parsing the text.