Add allAddTermsWith(arg)
Primarily used to skirt integration when taking partial derivatives
(see exact_equations_1).

If given a non-Add expression with the argument contained, it returns
the expression; otherwise, it returns an Add expression of all the
sub-expressions that contain the argument. That is, to get all
expressions in a Add function without a variable 'x' or 'y' (e.g.
"3x+4y".allAddTermsWith("x") = 3x).