Add generateMultivariateFunction
Generates a function of x and y out of power, log, and trig functions.
Similar to CalcFunctions, but this handles multiple variables and
doesn't need wrong answers.

If necessary, it can be generalized further at some later date and the
switch statement can be moved into the exercise(s) as needed to meet
their specific needs. This function was originally much more concise,
but it generated functions that were either trivial or messy to
differentiate.